refactor: auto-split test_conservation_laws.py
